+++ packages/tools/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./catalog.js";
+export * from "./filesystem.js";
 export * from "./policy.js";